backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.3
google-genai>=1.0.0
httpx>=0.26.0
reportlab>=4.1.0
python-dotenv>=1.0.0
networkx>=3.2.1
numpy>=2.0.0
pandas>=2.2.3
